Disect App.js into feed and submit
-Neue component - Dismiss Button
-Conditional Rendering in DisplayLink - onFeed als Boolean Argument
bestimmt of DismissButton und markAsRead Funktion angezeigt wird
